Add support for a simple event loop with proxy to the testing backend
Moved from the spawn_local test, this allows for re-use in the
show_strongref integration test.
