+++ api/rs/slint/Cargo.toml
@@ -167,7 +167,6 @@ log = { version = "0.4.17", optional = true }
 [dev-dependencies]
 slint-build = { path = "../build" }
 i-slint-backend-testing = { path = "../../../internal/backends/testing" }
-i-slint-backend-winit = { path = "../../../internal/backends/winit", features = ["wayland", "x11", "renderer-software"] }
 serde_json = "1.0.96"
 serde = { version = "1.0.163", features = ["derive"] }
